Menu Déroulant fonctionnel, ajout des liens et dimensions ainsi à voir (thiziri et binbing), et responsive à voir (hamza, thiziri et binbing...) !
