+++ app/src/main/res/layout-v21/activity_ecran_accueil.xml
@@ -10,6 +10,20 @@
     android:orientation="vertical"
     tools:context=".EcranAccueil">
 
+    <androidx.appcompat.widget.Toolbar
+        android:id="@+id/toolbar"
+        android:layout_height="wrap_content"
+        android:layout_width="match_parent"
+        android:background="#fff"
+        app:titleTextColor="@color/pink"
+        tools:ignore="MissingConstraints" />
+
+    <FrameLayout
+        android:id="@+id/container"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_weight="1"
+        />
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
