+++ app/src/main/res/layout/activity_main.xml
@@ -7,20 +7,7 @@
     tools:context=".MainActivity"
     android:background="@drawable/background">
 
-    <androidx.appcompat.widget.Toolbar
-        android:id="@+id/toolbar"
-        android:layout_height="wrap_content"
-        android:layout_width="match_parent"
-        android:background="#fff"
-        app:titleTextColor="@color/pink"
-        tools:ignore="MissingConstraints" />
 
-    <FrameLayout
-        android:id="@+id/container"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_weight="1"
-        />
     <ImageView
         android:id="@+id/imageAccueil"
         android:layout_width="518dp"
